requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
